added DataParser module containing necessary methods
